Address Timo's comments
