+++ flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/api/internal/InsertResultIterator.java
@@ -1,0 +1,71 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.api.internal;
+
+import org.apache.flink.annotation.Internal;
+import org.apache.flink.core.execution.JobClient;
+import org.apache.flink.table.api.TableException;
+import org.apache.flink.types.Row;
+import org.apache.flink.util.CloseableIterator;
+
+import java.util.NoSuchElementException;
+import java.util.Optional;
+
+/**
+ * A CloseableIterator for insert operation result.
+ */
+@Internal
+class InsertResultIterator implements CloseableIterator<Row> {
+	private final JobClient jobClient;
+	private final Row affectedRowCountsRow;
+	private Optional<Boolean> hasNext = Optional.empty();
+
+	InsertResultIterator(JobClient jobClient, Row affectedRowCountsRow) {
+		this.jobClient = jobClient;
+		this.affectedRowCountsRow = affectedRowCountsRow;
+	}
+
+	@Override
+	public void close() throws Exception {
+		jobClient.cancel();
+	}
+
+	@Override
+	public boolean hasNext() {
+		if (!hasNext.isPresent()) {
+			try {
+				jobClient.getJobExecutionResult(Thread.currentThread().getContextClassLoader()).get();
+			} catch (Exception e) {
+				throw new TableException("Failed to wait job finish", e);
+			}
+			hasNext = Optional.of(true);
+		}
+		return hasNext.get();
+	}
+
+	@Override
+	public Row next() {
+		if (hasNext()) {
+			hasNext = Optional.of(false);
+			return affectedRowCountsRow;
+		} else {
+			throw new NoSuchElementException();
+		}
+	}
+}
