+++ flink-table/flink-table-planner/src/main/scala/org/apache/flink/table/api/internal/TableEnvImpl.scala
@@ -44,7 +44,6 @@ import org.apache.flink.table.types.{AbstractDataType, DataType}
 import org.apache.flink.table.util.JavaScalaConversionUtil
 import org.apache.flink.table.utils.PrintUtils
 import org.apache.flink.types.Row
-import org.apache.flink.util.CloseableIterator
 
 import org.apache.calcite.jdbc.CalciteSchemaBuilder.asRootSchema
 import org.apache.calcite.sql.parser.SqlParser
@@ -53,7 +52,7 @@ import org.apache.commons.lang3.StringUtils
 
 import _root_.java.lang.{Iterable => JIterable, Long => JLong}
 import _root_.java.util.function.{Function => JFunction, Supplier => JSupplier}
-import _root_.java.util.{NoSuchElementException, Optional, Collections => JCollections, HashMap => JHashMap, List => JList, Map => JMap}
+import _root_.java.util.{Optional, Collections => JCollections, HashMap => JHashMap, List => JList, Map => JMap}
 
 import _root_.scala.collection.JavaConversions._
 import _root_.scala.collection.JavaConverters._
@@ -1283,38 +1282,5 @@ abstract class TableEnvImpl(
     val currentDatabase = catalogManager.getCurrentDatabase
 
     planningConfigurationBuilder.createFlinkPlanner(currentCatalogName, currentDatabase)
-  }
-
-  /**
-   * Iterator for insert operation result.
-   */
-  private final class InsertResultIterator(
-      jobClient: JobClient,
-      affectedRowCountsRow: Row) extends CloseableIterator[Row] {
-    private var hasNextRow: Option[Boolean] = Option.empty[Boolean]
-
-    override def close(): Unit = jobClient.cancel()
-
-    override def hasNext(): Boolean = {
-      if (hasNextRow.isEmpty) {
-        try {
-          jobClient.getJobExecutionResult(Thread.currentThread.getContextClassLoader).get
-        } catch {
-          case e: Exception => throw new TableException("Failed to wait job finish", e)
-        }
-        hasNextRow = Some(true)
-      }
-      hasNextRow.get
-    }
-
-    override def next(): Row = {
-      if (hasNext()) {
-        hasNextRow = Some(false)
-        affectedRowCountsRow
-      }
-      else {
-        throw new NoSuchElementException
-      }
-    }
   }
 }
